feat: add grading exam

+++ src/main/java/com/unimate/unimate/entity/UjianSiswa.java
@@ -28,5 +28,5 @@ public class UjianSiswa {
 
 
     @Column(name = "grade")
-    private Integer grade;
+    private Double grade;
 }
